FEM: Restructure python test modules/packages
There is a new split into app and gui tests. For this purpose two
packages where created: femtest.app and femtest.gui. Tests that don't
require the GUI up and running are inside the femtest.app package. Tests
that do require the GUI are inside the femtest.gui package. Also
underscores where added to the module names: testcommon to test_common.
