+++ src/Mod/Fem/CMakeLists.txt
@@ -130,66 +130,70 @@ SET(FemSolverZ88_SRCS
 
 SET(FemTests_SRCS
     femtest/__init__.py
-    femtest/testccxtools.py
-    femtest/testcommon.py
-    femtest/testfemimport.py
-    femtest/testmaterial.py
-    femtest/testmesh.py
-    femtest/testobject.py
-    femtest/testresult.py
-    femtest/testsolverframework.py
-    femtest/utilstest.py
+)
+
+SET(FemTestsApp_SRCS
+    femtest/app/__init__.py
+    femtest/app/support_utils.py
+    femtest/app/test_ccxtools.py
+    femtest/app/test_common.py
+    femtest/app/test_femimport.py
+    femtest/app/test_material.py
+    femtest/app/test_mesh.py
+    femtest/app/test_object.py
+    femtest/app/test_result.py
+    femtest/app/test_solverframework.py
 )
 
 SET(FemTestsFiles_SRCS
-    femtest/testfiles/__init__.py
+    femtest/data/__init__.py
 )
 
 SET(FemTestsCcx_SRCS
-    femtest/testfiles/ccx/__init__.py
-    femtest/testfiles/ccx/cube_mesh.py
-    femtest/testfiles/ccx/cube_frequency.inp
-    femtest/testfiles/ccx/cube_frequency.dat
-    femtest/testfiles/ccx/cube_frequency.frd
-    femtest/testfiles/ccx/cube_frequency_expected_values
-    femtest/testfiles/ccx/cube_frequency.FCStd
-    femtest/testfiles/ccx/cube_static.inp
-    femtest/testfiles/ccx/cube_static.dat
-    femtest/testfiles/ccx/cube_static.frd
-    femtest/testfiles/ccx/cube_static_expected_values
-    femtest/testfiles/ccx/cube_static.FCStd
-    femtest/testfiles/ccx/cube.FCStd
-    femtest/testfiles/ccx/multimat.inp
-    femtest/testfiles/ccx/multimat_mesh.py
-    femtest/testfiles/ccx/spine_mesh.py
-    femtest/testfiles/ccx/spine_thermomech.inp
-    femtest/testfiles/ccx/spine_thermomech.dat
-    femtest/testfiles/ccx/spine_thermomech.frd
-    femtest/testfiles/ccx/spine_thermomech_expected_values
-    femtest/testfiles/ccx/spine_thermomech.FCStd
-    femtest/testfiles/ccx/Flow1D_mesh.py
-    femtest/testfiles/ccx/Flow1D_thermomech.inp
-    femtest/testfiles/ccx/Flow1D_thermomech.dat
-    femtest/testfiles/ccx/Flow1D_thermomech.frd
-    femtest/testfiles/ccx/Flow1D_thermomech_expected_values
-    femtest/testfiles/ccx/Flow1D_thermomech_inout_nodes.txt
-    femtest/testfiles/ccx/Flow1D_thermomech.FCStd
+    femtest/data/ccx/__init__.py
+    femtest/data/ccx/cube_mesh.py
+    femtest/data/ccx/cube_frequency.inp
+    femtest/data/ccx/cube_frequency.dat
+    femtest/data/ccx/cube_frequency.frd
+    femtest/data/ccx/cube_frequency_expected_values
+    femtest/data/ccx/cube_frequency.FCStd
+    femtest/data/ccx/cube_static.inp
+    femtest/data/ccx/cube_static.dat
+    femtest/data/ccx/cube_static.frd
+    femtest/data/ccx/cube_static_expected_values
+    femtest/data/ccx/cube_static.FCStd
+    femtest/data/ccx/cube.FCStd
+    femtest/data/ccx/multimat.inp
+    femtest/data/ccx/multimat_mesh.py
+    femtest/data/ccx/spine_mesh.py
+    femtest/data/ccx/spine_thermomech.inp
+    femtest/data/ccx/spine_thermomech.dat
+    femtest/data/ccx/spine_thermomech.frd
+    femtest/data/ccx/spine_thermomech_expected_values
+    femtest/data/ccx/spine_thermomech.FCStd
+    femtest/data/ccx/Flow1D_mesh.py
+    femtest/data/ccx/Flow1D_thermomech.inp
+    femtest/data/ccx/Flow1D_thermomech.dat
+    femtest/data/ccx/Flow1D_thermomech.frd
+    femtest/data/ccx/Flow1D_thermomech_expected_values
+    femtest/data/ccx/Flow1D_thermomech_inout_nodes.txt
+    femtest/data/ccx/Flow1D_thermomech.FCStd
 )
 
 SET(FemTestsElmer_SRCS
-    femtest/testfiles/elmer/__init__.py
-    femtest/testfiles/elmer/case.sif
-    femtest/testfiles/elmer/group_mesh.geo
-    femtest/testfiles/elmer/ELMERSOLVER_STARTINFO
+    femtest/data/elmer/__init__.py
+    femtest/data/elmer/case.sif
+    femtest/data/elmer/group_mesh.geo
+    femtest/data/elmer/ELMERSOLVER_STARTINFO
 )
 
 SET(FemTestsMesh_SRCS
-    femtest/testfiles/mesh/__init__.py
-    femtest/testfiles/mesh/tetra10_mesh.inp
-    femtest/testfiles/mesh/tetra10_mesh.unv
-    femtest/testfiles/mesh/tetra10_mesh.vtk
-    femtest/testfiles/mesh/tetra10_mesh.yml
-    femtest/testfiles/mesh/tetra10_mesh.z88
+    femtest/data/mesh/__init__.py
+    femtest/data/mesh/tetra10_mesh.inp
+    femtest/data/mesh/tetra10_mesh.unv
+    femtest/data/mesh/tetra10_mesh.vtk
+    femtest/data/mesh/tetra10_mesh.yml
+    femtest/data/mesh/tetra10_mesh.z88
 )
 
 SET(FemTools_SRCS
@@ -237,6 +241,7 @@ SET(FemAllScripts
     ${FemSolverFenics_SRCS}
     ${FemSolverZ88_SRCS}
     ${FemTests_SRCS}
+    ${FemTestsApp_SRCS}
     ${FemTestsFiles_SRCS}
     ${FemTestsCcx_SRCS}
     ${FemTestsElmer_SRCS}
@@ -268,10 +273,11 @@ INSTALL(FILES ${FemSolverElmerEquations_SRCS} DESTINATION Mod/Fem/femsolver/elme
 INSTALL(FILES ${FemSolverFenics_SRCS} DESTINATION Mod/Fem/femsolver/fenics)
 INSTALL(FILES ${FemSolverZ88_SRCS} DESTINATION Mod/Fem/femsolver/z88)
 INSTALL(FILES ${FemTests_SRCS} DESTINATION Mod/Fem/femtest)
-INSTALL(FILES ${FemTestsFiles_SRCS} DESTINATION Mod/Fem/femtest/testfiles)
-INSTALL(FILES ${FemTestsCcx_SRCS} DESTINATION Mod/Fem/femtest/testfiles/ccx)
-INSTALL(FILES ${FemTestsElmer_SRCS} DESTINATION Mod/Fem/femtest/testfiles/elmer)
-INSTALL(FILES ${FemTestsMesh_SRCS} DESTINATION Mod/Fem/femtest/testfiles/mesh)
+INSTALL(FILES ${FemTestsApp_SRCS} DESTINATION Mod/Fem/femtest/app)
+INSTALL(FILES ${FemTestsFiles_SRCS} DESTINATION Mod/Fem/femtest/data)
+INSTALL(FILES ${FemTestsCcx_SRCS} DESTINATION Mod/Fem/femtest/data/ccx)
+INSTALL(FILES ${FemTestsElmer_SRCS} DESTINATION Mod/Fem/femtest/data/elmer)
+INSTALL(FILES ${FemTestsMesh_SRCS} DESTINATION Mod/Fem/femtest/data/mesh)
 INSTALL(FILES ${FemTools_SRCS} DESTINATION Mod/Fem/femtools)
 INSTALL(FILES ${FemObjectsScripts_SRCS} DESTINATION Mod/Fem/femobjects)
