+++ src/Mod/Fem/TestFem.py
@@ -26,16 +26,16 @@
 
 # Unit test for the FEM module
 # to get the right order import as is used
-from femtest.testfemimport import TestFemImport as FemTest01
-from femtest.testcommon import TestFemCommon as FemTest02
-from femtest.testobject import TestObjectCreate as FemTest03
-from femtest.testobject import TestObjectType as FemTest04
-from femtest.testmaterial import TestMaterialUnits as FemTest05
-from femtest.testmesh import TestMeshCommon as FemTest06
-from femtest.testmesh import TestMeshEleTetra10 as FemTest07
-from femtest.testresult import TestResult as FemTest08
-from femtest.testccxtools import TestCcxTools as FemTest09
-from femtest.testsolverframework import TestSolverFrameWork as FemTest10
+from femtest.app.test_femimport import TestFemImport as FemTest01
+from femtest.app.test_common import TestFemCommon as FemTest02
+from femtest.app.test_object import TestObjectCreate as FemTest03
+from femtest.app.test_object import TestObjectType as FemTest04
+from femtest.app.test_material import TestMaterialUnits as FemTest05
+from femtest.app.test_mesh import TestMeshCommon as FemTest06
+from femtest.app.test_mesh import TestMeshEleTetra10 as FemTest07
+from femtest.app.test_result import TestResult as FemTest08
+from femtest.app.test_ccxtools import TestCcxTools as FemTest09
+from femtest.app.test_solverframework import TestSolverFrameWork as FemTest10
 
 # dummy usage to get flake8 and lgtm quiet
 False if FemTest01.__name__ else True
@@ -258,13 +258,12 @@ unittest.TextTestRunner().run(unittest.TestLoader().loadTestsFromName("femtest.t
 # open files from FEM test suite source code
 # be careful on updating these files, they contain the original results!
 # TODO update files, because some of them have non-existing FEM object classes
-app_home = FreeCAD.ConfigGet("AppHomePath")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/cube.FCStd")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/cube_frequency.FCStd")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/cube_static.FCStd")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/Flow1D_thermomech.FCStd")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/multimat.FCStd")
-doc = FreeCAD.open(app_home + "Mod/Fem/femtest/testfiles/ccx/spine_thermomech.FCStd")
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/cube.FCStd')
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/cube_frequency.FCStd')
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/cube_static.FCStd')
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/Flow1D_thermomech.FCStd')
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/multimat.FCStd')
+doc = FreeCAD.open(FreeCAD.ConfigGet("AppHomePath") + 'Mod/Fem/femtest/data/ccx/spine_thermomech.FCStd')
 
 # open files generated from test suite
 import femtest.utilstest as ut
