+++ src/Mod/Fem/femtest/app/test_result.py
@@ -25,8 +25,8 @@
 
 import FreeCAD
 import unittest
-from . import utilstest as testtools
-from .utilstest import fcc_print
+from . import support_utils as testtools
+from .support_utils import fcc_print
 
 from os.path import join
