+++ src/Mod/Fem/femtest/data/ccx/cube.FCStd
