+++ src/Mod/Fem/femtest/data/ccx/cube_frequency.inp
