+++ src/Mod/Fem/femtest/data/ccx/cube_mesh.py
