+++ src/Mod/Fem/femtest/data/ccx/cube_static.frd
