+++ src/Mod/Fem/femtest/data/ccx/cube_static_expected_values
