+++ src/Mod/Fem/femtest/data/ccx/spine_mesh.py
