+++ src/Mod/Fem/femtest/data/ccx/spine_points.csv
