+++ src/Mod/Fem/femtest/data/ccx/spine_thermomech.FCStd
