+++ src/Mod/Fem/femtest/data/ccx/spine_thermomech_expected_values
