+++ src/Mod/Fem/femtest/data/mesh/tetra10_mesh.inp
